+++ src/commonMain/kotlin/at/posselt/pfrpg2e/kingdom/modifiers/ModifierSelector.kt
@@ -1,0 +1,20 @@
+package at.posselt.pfrpg2e.kingdom.modifiers
+
+import at.posselt.pfrpg2e.data.ValueEnum
+import at.posselt.pfrpg2e.fromCamelCase
+import at.posselt.pfrpg2e.localization.Translatable
+import at.posselt.pfrpg2e.toCamelCase
+
+enum class ModifierSelector : Translatable, ValueEnum {
+    CHECK,
+    CONSUMPTION;
+
+    companion object {
+        fun fromString(value: String) = fromCamelCase<ModifierSelector>(value)
+    }
+
+    override val value: String
+        get() = toCamelCase()
+
+    override val i18nKey = "modifierSelector.$value"
+}