+++ src/commonMain/kotlin/at/posselt/pfrpg2e/kingdom/modifiers/evaluation/FilterModifiers.kt
@@ -1,6 +1,7 @@
 package at.posselt.pfrpg2e.kingdom.modifiers.evaluation
 
 import at.posselt.pfrpg2e.kingdom.modifiers.Modifier
+import at.posselt.pfrpg2e.kingdom.modifiers.ModifierSelector
 import at.posselt.pfrpg2e.kingdom.modifiers.expressions.ExpressionContext
 
 data class FilterResult(
@@ -11,13 +12,16 @@ data class FilterResult(
 fun filterModifiersAndUpdateContext(
     modifiers: List<Modifier>,
     context: ExpressionContext,
+    selector: ModifierSelector,
 ): FilterResult {
+    // only keep modifiers that match selector
+    val selectedMods = modifiers.filter { it.selector == selector }
     // first filter applicable modifiers and get their roll options
-    val rollOptions = modifiers.asSequence()
+    val rollOptions = selectedMods.asSequence()
         .filter { it.applyIf.all { it.evaluate(context) } && it.enabled }
         .flatMap { it.rollOptions }
         .toSet()
-    val filteredModifiers = modifiers.asSequence()
+    val filteredModifiers = selectedMods.asSequence()
         .filter { it.applyIf.all { it.evaluate(context.copy(rollOptions = context.rollOptions + rollOptions)) } }
         .toList()
     val enabledRollOptions = filteredModifiers.flatMap { it.rollOptions }.toSet()