+++ src/commonMain/kotlin/at/posselt/pfrpg2e/kingdom/resources/Consumption.kt
@@ -2,6 +2,11 @@ package at.posselt.pfrpg2e.kingdom.resources
 
 import at.posselt.pfrpg2e.data.kingdom.RealmData
 import at.posselt.pfrpg2e.data.kingdom.settlements.Settlement
+import at.posselt.pfrpg2e.kingdom.modifiers.Modifier
+import at.posselt.pfrpg2e.kingdom.modifiers.ModifierSelector
+import at.posselt.pfrpg2e.kingdom.modifiers.evaluation.evaluateModifiers
+import at.posselt.pfrpg2e.kingdom.modifiers.evaluation.filterModifiersAndUpdateContext
+import at.posselt.pfrpg2e.kingdom.modifiers.expressions.ExpressionContext
 
 data class Consumption(
     val now: Int,
@@ -12,6 +17,7 @@ data class Consumption(
     val armies: Int,
     val settlements: Int,
     val resourcesSurplus: Int,
+    val modifierConsumption: Int,
 )
 
 fun calculateConsumption(
@@ -19,12 +25,17 @@ fun calculateConsumption(
     realmData: RealmData,
     armyConsumption: Int,
     now: Int,
+    expressionContext: ExpressionContext,
+    modifiers: List<Modifier>,
 ): Consumption {
+    val modifiers = filterModifiersAndUpdateContext(modifiers, expressionContext, ModifierSelector.CONSUMPTION)
+    val result = evaluateModifiers(modifiers)
+    val modifierConsumption = result.total
     val farmlands = realmData.worksites.farmlands.quantity
     val food = realmData.worksites.farmlands.resources
     val settlementConsumption = settlements.sumOf { it.consumption }
     val settlementSurplus = settlements.sumOf { it.consumptionSurplus }
-    val consumers = now + settlementConsumption + armyConsumption
+    val consumers = now + settlementConsumption + armyConsumption + modifierConsumption
     val total = (consumers - food - farmlands).coerceIn(0, Int.MAX_VALUE)
     val resourcesSurplus = (farmlands + food - consumers).coerceIn(0, Int.MAX_VALUE)
     return Consumption(
@@ -36,5 +47,6 @@ fun calculateConsumption(
         armies = armyConsumption,
         settlements = settlementConsumption,
         now = now,
+        modifierConsumption = modifierConsumption,
     )
 }