+++ src/jsMain/kotlin/at/posselt/pfrpg2e/kingdom/sheet/contexts/ConsumptionBreakdownContext.kt
@@ -2,6 +2,7 @@ package at.posselt.pfrpg2e.kingdom.sheet.contexts
 
 import at.posselt.pfrpg2e.kingdom.resources.Consumption
 import js.objects.JsPlainObject
+import kotlin.math.abs
 
 @Suppress("unused")
 @JsPlainObject
@@ -14,6 +15,8 @@ external interface ConsumptionBreakdownContext {
     val settlements: Int
     val resourcesSurplus: Int
     val now: Int
+    val modifiers: Int
+    val modifiersSign: String
 }
 
 fun Consumption.toContext() = ConsumptionBreakdownContext(
@@ -25,4 +28,6 @@ fun Consumption.toContext() = ConsumptionBreakdownContext(
     settlements = settlements,
     resourcesSurplus = resourcesSurplus,
     now = now,
+    modifiers = abs(modifierConsumption),
+    modifiersSign = if (modifierConsumption < 0) "-" else "+"
 )