+++ src/jsMain/kotlin/at/posselt/pfrpg2e/kingdom/sheet/contexts/ModifierContext.kt
@@ -1,6 +1,7 @@
 package at.posselt.pfrpg2e.kingdom.sheet.contexts
 
 import at.posselt.pfrpg2e.kingdom.RawModifier
+import at.posselt.pfrpg2e.kingdom.modifiers.ModifierSelector
 import at.posselt.pfrpg2e.utils.t
 import js.objects.JsPlainObject
 
@@ -10,6 +11,7 @@ external interface ModifierContext {
     val name: String
     val description: String
     val turns: String
+    val selector: String
     val isConsumedAfterRoll: Boolean
 }
 
@@ -21,6 +23,9 @@ fun Array<RawModifier>.toContext(): Array<ModifierContext> {
         val label = if(it.requiresTranslation != false && buttonLabel != null) t(buttonLabel) else buttonLabel
         ModifierContext(
             name = name,
+            selector = t(it.selector
+                ?.let { value -> ModifierSelector.fromString(value) }
+                ?: ModifierSelector.CHECK),
             description = label ?: name,
             turns = if ((it.turns ?: 0) > 0) "${it.turns}" else t("kingdom.indefinite"),
             isConsumedAfterRoll = it.isConsumedAfterRoll == true,