+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 * The module can now be translated into different languages
+* Add modifiers that increase consumption
+* Automate Squatters event consumption increase
 
 ### Changed
 * Require FoundryVTT 13